+++ docs/changelog.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## WIP
+
+* When plotting a histogram against a PDF or other histogram with
+  `plot_pull` or `plot_ratio`, normalize the histogram values to the mean bin
+  width to correct for irregular bin widths when using axes of type
+  `axis.Variable`.
+  [#369](https://github.com/scikit-hep/hist/pull/369)
+
 ## Version 2.6.1
 
 * Fall back on normal repr when histogram is too large